adjusted createArray script to output exactly what had been handwritten and moved certain functions from additions.h and additions.c to actual implementation functions inside the respective classes. reference counter and releaser. copyArray funcs..they still need reworking.

+++ src/ExprArray.c
@@ -8,9 +8,14 @@ void init_ExprArray(ExprArray* array){
         array->Exprs = NULL;
         array->size = 0;
         array->used = 0;
+        array->owner_references =1;
         array->addElementToArray = &addElementToExprArray;
         array->deleteArray = &delete_ExprArray;
         array->getElementInArrayAt =&getExprinArrayAt;
+        array->deleteArray = &delete_ExprArray;
+        array->getArrayReference = &getExprArrayReference;
+        array->releaseArrayReference = &releaseExprArrayReference;
+        array->copyArray = &copyExprArray;
     }
             
 void addElementToExprArray(ExprArray* array,Expr* element){
@@ -35,8 +40,13 @@ void addElementToExprArray(ExprArray* array,Expr* element){
 void delete_ExprArray(ExprArray* array){
     if(array){
         int i;
-        for(i =0; i <array->used;i++){
-            delete_Expr(array->Exprs[i]);
+        if(array->owner_references<=1){
+            for(i =0; i <array->used;i++){
+                delete_Expr(array->Exprs[i]);
+            }
+        }
+        else{
+            releaseExprArrayReference(array);
         }
     }
 }
@@ -45,3 +55,23 @@ Expr* getExprinArrayAt(ExprArray* array,size_t index){
         return array->Exprs[index];
     return NULL;
 }
+ExprArray*  copyExprArray(ExprArray * arr){
+    ExprArray* newarr;
+    int i;
+    newarr = malloc(sizeof(ExprArray));
+    init_ExprArray(newarr);
+    for(i=0;i<arr->used;i++){
+        newarr->addElementToArray(newarr, getExprinArrayAt(arr,i));
+    }
+    return newarr;
+}
+ExprArray* getExprArrayReference(ExprArray* arr){
+    arr->owner_references++;
+    return arr;
+}
+ExprArray* releaseExprArrayReference(ExprArray* arr){
+    arr->owner_references--;
+    return NULL;
+}
+
+
